+++ packages/astro/src/core/config/config.ts
@@ -6,6 +6,7 @@ import * as colors from 'kleur/colors';
 import path from 'path';
 import { fileURLToPath, pathToFileURL } from 'url';
 import { mergeConfig as mergeViteConfig } from 'vite';
+import { isHybridMalconfigured } from '../../prerender/utils.js';
 import { AstroError, AstroErrorData } from '../errors/index.js';
 import type { LogOptions } from '../logger/core.js';
 import { arraify, isObject, isURL } from '../util.js';
@@ -103,6 +104,8 @@ export function resolveFlags(flags: Partial<Flags>): CLIFlags {
 		drafts: typeof flags.drafts === 'boolean' ? flags.drafts : undefined,
 		experimentalAssets:
 			typeof flags.experimentalAssets === 'boolean' ? flags.experimentalAssets : undefined,
+		experimentalMiddleware:
+			typeof flags.experimentalMiddleware === 'boolean' ? flags.experimentalMiddleware : undefined,
 	};
 }
 
@@ -135,6 +138,9 @@ function mergeCLIFlags(astroConfig: AstroUserConfig, flags: CLIFlags) {
 		// @ts-expect-error astroConfig.server may be a function, but TS doesn't like attaching properties to a function.
 		// TODO: Come back here and refactor to remove this expected error.
 		astroConfig.server.open = flags.open;
+	}
+	if (typeof flags.experimentalMiddleware === 'boolean') {
+		astroConfig.experimental.middleware = true;
 	}
 	return astroConfig;
 }
@@ -217,6 +223,12 @@ export async function openConfig(configOptions: LoadConfigOptions): Promise<Open
 		userConfig = config.value;
 	}
 	const astroConfig = await resolveConfig(userConfig, root, flags, configOptions.cmd);
+
+	if (isHybridMalconfigured(astroConfig)) {
+		throw new Error(
+			`The "output" config option must be set to "hybrid" and "experimental.hybridOutput" must be set to true to use the hybrid output mode. Falling back to "static" output mode.`
+		);
+	}
 
 	return {
 		astroConfig,