+++ packages/astro/src/core/config/vite-load.ts
@@ -11,7 +11,7 @@ export interface ViteLoader {
 async function createViteLoader(root: string, fs: typeof fsType): Promise<ViteLoader> {
 	const viteServer = await vite.createServer({
 		server: { middlewareMode: true, hmr: false },
-		optimizeDeps: { entries: [] },
+		optimizeDeps: { disabled: true },
 		clearScreen: false,
 		appType: 'custom',
 		ssr: {
@@ -24,6 +24,7 @@ async function createViteLoader(root: string, fs: typeof fsType): Promise<ViteLo
 				'@astrojs/react',
 				'@astrojs/preact',
 				'@astrojs/sitemap',
+				'@astrojs/markdoc',
 			],
 		},
 		plugins: [loadFallbackPlugin({ fs, root: pathToFileURL(root) })],