+++ packages/astro/test/astro-slot-with-client.test.js
@@ -16,4 +16,10 @@ describe('Slots with client: directives', () => {
 		const $ = cheerio.load(html);
 		expect($('script')).to.have.a.lengthOf(1);
 	});
+
+	it('Astro slot tags are cleaned', async () => {
+		const html = await fixture.readFile('/index.html');
+		const $ = cheerio.load(html);
+		expect($('astro-slot')).to.have.a.lengthOf(0);
+	});
 });