+++ packages/integrations/netlify/test/edge-functions/root-dynamic.test.ts
@@ -1,20 +1,19 @@
-// @ts-expect-error
-import { runBuild } from './test-utils.ts';
-// @ts-expect-error
+import { loadFixture } from './test-utils.ts';
 import { assertEquals, assert, DOMParser } from './deps.ts';
 
-// @ts-expect-error
 Deno.test({
 	// TODO: debug why build cannot be found in "await import"
 	ignore: true,
 	name: 'Assets are preferred over HTML routes',
 	async fn() {
-		let close = await runBuild('./fixtures/root-dynamic/');
+		const fixture = loadFixture('./fixtures/root-dynamic/');
+		await fixture.runBuild();
+
 		const { default: handler } = await import(
 			'./fixtures/root-dynamic/.netlify/edge-functions/entry.js'
 		);
 		const response = await handler(new Request('http://example.com/styles.css'));
 		assertEquals(response, undefined, 'No response because this is an asset');
-		await close();
+		await fixture.cleanup();
 	},
 });