+++ packages/integrations/vercel/src/static/adapter.ts
@@ -1,6 +1,14 @@
 import type { AstroAdapter, AstroConfig, AstroIntegration } from 'astro';
 
+import {
+	defaultImageConfig,
+	getImageConfig,
+	throwIfAssetsNotEnabled,
+	type VercelImageConfig,
+} from '../image/shared.js';
+import { exposeEnv } from '../lib/env.js';
 import { emptyDir, getVercelOutput, writeJson } from '../lib/fs.js';
+import { isHybridOutput } from '../lib/prerender.js';
 import { getRedirects } from '../lib/redirects.js';
 
 const PACKAGE_NAME = '@astrojs/vercel/static';
@@ -11,26 +19,43 @@ function getAdapter(): AstroAdapter {
 
 export interface VercelStaticConfig {
 	analytics?: boolean;
+	imageService?: boolean;
+	imagesConfig?: VercelImageConfig;
 }
 
-export default function vercelStatic({ analytics }: VercelStaticConfig = {}): AstroIntegration {
+export default function vercelStatic({
+	analytics,
+	imageService,
+	imagesConfig,
+}: VercelStaticConfig = {}): AstroIntegration {
 	let _config: AstroConfig;
 
 	return {
 		name: '@astrojs/vercel',
 		hooks: {
-			'astro:config:setup': ({ command, config, injectScript }) => {
+			'astro:config:setup': ({ command, config, injectScript, updateConfig }) => {
 				if (command === 'build' && analytics) {
 					injectScript('page', 'import "@astrojs/vercel/analytics"');
 				}
-				config.outDir = new URL('./static/', getVercelOutput(config.root));
-				config.build.format = 'directory';
+				const outDir = new URL('./static/', getVercelOutput(config.root));
+				const viteDefine = exposeEnv(['VERCEL_ANALYTICS_ID']);
+				updateConfig({
+					outDir,
+					build: {
+						format: 'directory',
+					},
+					vite: {
+						define: viteDefine,
+					},
+					...getImageConfig(imageService, imagesConfig, command),
+				});
 			},
 			'astro:config:done': ({ setAdapter, config }) => {
+				throwIfAssetsNotEnabled(config, imageService);
 				setAdapter(getAdapter());
 				_config = config;
 
-				if (config.output === 'server') {
+				if (config.output === 'server' || isHybridOutput(config)) {
 					throw new Error(`${PACKAGE_NAME} should be used with output: 'static'`);
 				}
 			},
@@ -46,6 +71,9 @@ export default function vercelStatic({ analytics }: VercelStaticConfig = {}): As
 				await writeJson(new URL(`./config.json`, getVercelOutput(_config.root)), {
 					version: 3,
 					routes: [...getRedirects(routes, _config), { handle: 'filesystem' }],
+					...(imageService || imagesConfig
+						? { images: imagesConfig ? imagesConfig : defaultImageConfig }
+						: {}),
 				});
 			},
 		},