+++ packages/integrations/vercel/test/serverless-prerender.test.js
@@ -1,11 +1,12 @@
-import { loadFixture } from './test-utils.js';
 import { expect } from 'chai';
+import { loadFixture } from './test-utils.js';
 
 describe('Serverless prerender', () => {
 	/** @type {import('./test-utils').Fixture} */
 	let fixture;
 
 	before(async () => {
+		process.env.PRERENDER = true;
 		fixture = await loadFixture({
 			root: './fixtures/serverless-prerender/',
 		});
@@ -13,6 +14,27 @@ describe('Serverless prerender', () => {
 
 	it('build successful', async () => {
 		await fixture.build();
-		expect(fixture.readFile('/static/index.html')).to.be.ok;
+		expect(await fixture.readFile('../.vercel/output/static/index.html')).to.be.ok;
+	});
+});
+
+describe('Serverless hybrid rendering', () => {
+	/** @type {import('./test-utils').Fixture} */
+	let fixture;
+
+	before(async () => {
+		process.env.PRERENDER = true;
+		fixture = await loadFixture({
+			root: './fixtures/serverless-prerender/',
+			output: 'hybrid',
+			experimental: {
+				hybridOutput: true,
+			},
+		});
+	});
+
+	it('build successful', async () => {
+		await fixture.build();
+		expect(await fixture.readFile('../.vercel/output/static/index.html')).to.be.ok;
 	});
 });